Fix OpCode::execute() dst out of bounds
We currently just check if the given start index and the following
indices needed for execution are in bounds.
But we take the destination index to look up the value and use this
value as index to write the result to.
We never check this position, so writing may panic.

Check each index respectively.